backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Mechanical digital library backend api"
authors = ["reechee <richardokonicha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
itsdangerous = "^1.1.0"
Authlib = "^0.15.3"
httpx = "^0.17.1"
motor = "^2.4.0"
python-decouple = "^3.4"
dnspython = "^2.1.0"
email-validator = "^1.1.2"
jwt = "^1.2.0"
PyJWT = "^2.0.1"
python-multipart = "^0.0.5"
mkdocs-material = "^7.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
